function pointers array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,4 @@
+#include <stdlib.h>
 #include "function_pointers.h"
 
 /**
@@ -15,7 +16,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 	{
 		return;
 	}
-	size_t i;
+	int i;
 
 	for (i = 0; i < size; i++)
 	{
